Updated readme with more detail and use my project.
The code by default uses my  project 'inux-academy-project-91522' instead of
'eng-lightning-244220'.The  intention is that a new user can set up
there own google project and test this code before merging.I have tested
this on my machine. I  expect that the google Auth steps for the APIs
need to  described in more detail. The requirements file probably
contains some junk, I  copied and  pasted it from the python virtual env
I  used for testing this but  did not  clean it up.

The readme  is intended to  contain enough  information for  someone
familiar with google  cloud to  run these cloud functions.   In the
future the scraper module must be expanded to scrape  other  sources.The
URLS filters and article data monging need customising  for  news
source. The architecture we will use  is not decided.  ATM, the  URLs are
published to a topic and   them the topic is subscribed to by another
function which publishes to big query. I made this cloud functions PoC  because
they  are  lightweight. Many sources are already scraped by  containers.
They are more portable between cloud envs.

Signed-off-by: Richard Hill <ricjhill@protonmail.com>